requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
streamlit==1.50.0
requests==2.32.3
python-dotenv==1.0.1
